requirements.txt
pyautogui
requests
Pillow
